backend/requirements.txt
# Azure Services
azure-identity
azure-keyvault-secrets
azure-ai-projects
azure-ai-agents
azure-storage-blob
azure-cosmos
azure-search-documents

# AI + Azure
openai
# azure-ai-ml removed (not used in code, was pulling large dependency graph causing build timeouts)
azure-core

# MCP Server Framework
mcp

# Diagram generation
graphviz
diagrams

# Optional, if using GraphViz locally
# pygraphviz

# For image handling / encoding (optional)
pillow

# Web Framework & API
fastapi
uvicorn
python-multipart
pydantic
httpx

# Utilities
python-dotenv